refactored fetch and aggregate candles. changed operator name to task
