--- new_pipeline/metadata.json
@@ -1,11 +1,0 @@
-{
-    "name": "Candlestick dataflow pipeline",
-    "description": "A Dataflow pipeline for fetching candlestick data and applying transformations",
-    "parameters": [
-        {
-            "name": "write_dest",
-            "label": "Write destination",
-            "helpText": "Write destination of pipeline output e.g. text or bigquery"
-        }
-    ]
-}
