--- new_pipeline/utils/pipeline_utils.py
@@ -1,8 +1,0 @@
-class Transformer:
-    l1 = []
-
-x = Transformer()
-x.l1 = 'hello'
-print(x.l1)
-
-print('a' >> 'b')
