+++ trading/cloudbuild.yaml
@@ -1,0 +1,25 @@
+steps:
+  - name: 'gcr.io/cloud-builders/docker'
+    id: 'build_docker_image'
+    args: ['build', '-t', 'us-central1-docker.pkg.dev/crypto-trading-v2/dataflow/pipeline', '-f', 'pipeline/Dockerfile', '.'] 
+
+  - name: 'gcr.io/cloud-builders/gcloud'
+    id: 'create_flex_template'
+    args: [
+      'dataflow', 'flex-template', 'build', 'gs://crypto-trading-v2-dataflow/pipeline.json', 
+      '--image', 'us-central1-docker.pkg.dev/crypto-trading-v2/dataflow/pipeline',
+      '--sdk-language', 'PYTHON',
+      '--metadata-file', 'pipeline/metadata.json'
+    ]
+    waitFor: ['build_docker_image']
+  - name: 'gcr.io/cloud-builders/gcloud'
+    id: 'run_flex_template'
+    args: [
+      'dataflow', 'flex-template', 'run', 'pipeline',
+      '--template-file-gcs-location', 'gs://crypto-trading-v2-dataflow/pipeline.json',
+      '--parameters', 'write_dest=bigquery'
+    ]
+    waitFor: ['create_flex_template']
+images:
+- 'us-central1-docker.pkg.dev/crypto-trading-v2/dataflow/pipeline'
+
