+++ trading/utils/pipeline_utils.py
@@ -1,0 +1,8 @@
+class Transformer:
+    l1 = []
+
+x = Transformer()
+x.l1 = 'hello'
+print(x.l1)
+
+print('a' >> 'b')
